# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "\u9879\u76ee\u7ba1\u7406\u7cfb\u7edf",
        "version": "1.0.0"
    },
    "host": "pmsapi.turtletl.com",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "\u7528\u6237\u767b\u5f55",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "\u767b\u5f55\u90ae\u7bb1",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "\u767b\u5f55\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "\u7528\u6237\u6ce8\u518c",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "\u7528\u6237\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "\u767b\u5f55\u90ae\u7bb1",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "\u767b\u5f55\u5bc6\u7801",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "register success"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/company/add": {
            "post": {
                "tags": [
                    "Companys"
                ],
                "summary": "\u65b0\u5efa\u516c\u53f8",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "\u7528\u6237\u51ed\u8bc1",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "company_name",
                        "in": "query",
                        "description": ">\u516c\u53f8\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "english_name",
                        "in": "query",
                        "description": ">\u82f1\u6587\u540d",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "company_code",
                        "in": "query",
                        "description": ">\u516c\u53f8\u4ee3\u53f7",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add successful"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/users/all": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "\u663e\u793a\u6240\u6709\u7528\u6237",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "\u7528\u6237\u51ed\u8bc1",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "all users"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        },
        "/users/one": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "\u83b7\u53d6\u5f53\u524d\u7528\u6237",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "\u7528\u6237\u51ed\u8bc1",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "one user"
                    },
                    "default": {
                        "description": "an \"unexpected\" error"
                    }
                }
            }
        }
    },
    "definitions": {},
    "tags": [
        {
            "name": "Auth",
            "description": "\u9a8c\u8bc1\u6a21\u5757"
        },
        {
            "name": "Users",
            "description": "\u7528\u6237\u6a21\u5757"
        },
        {
            "name": "Companys",
            "description": "\u516c\u53f8\u6a21\u5757"
        }
    ]
}